Remove ClubsPage.jsx
